requirements.txt
streamlit
pillow
opencv-python
scikit-image
google-generativeai